updating everything to uint32 and adding some bound checking on iteration increments

+++ ops/migrations/000002_hdwallet.up.sql
@@ -1,1 +1,1 @@
-ALTER TABLE keys ADD COLUMN iteration INTEGER;
+ALTER TABLE keys ADD COLUMN iteration BIGINT;
